[FIX] ld    %4294902082,r3 int overflow

+++ include/robot_factory.h
@@ -55,7 +55,7 @@ int compile_line(op_t *operation, arg_t *args, assm_cfg_t *assm_cfg);
 int robot_factory(int argc, char **argv);
 int parse_file(char *file_path, assm_cfg_t *assm_cfg);
 int write_byte(char byte, assm_cfg_t *assm_cfg);
-void write_bytes(int num, int byte_nb, assm_cfg_t *assm_cfg);
+void write_bytes(long num, int byte_nb, assm_cfg_t *assm_cfg);
 int write_to_header(char *str, assm_cfg_t *assm_cfg, raw_str_type_e type);
 int write_buff_to_output(assm_cfg_t *assm_cfg);
 int write_header_to_output(assm_cfg_t *assm_cfg);
